+++ examples/ScnLoaderExample.py
@@ -2,17 +2,20 @@ import os
 import json
 import sys
 sys.path.append("..")
-from src.tools.DataLoader import Scenes,Scnfolder,Select
+from src.tools.DataLoader import Scenes,Scnfolder,Select,Config
+
+#Config.version = 'sanoba witch'
 
 file_path = "D:\\games\\senrenbanka\\outPath\\data.xp3\\scn\\001・アーサー王ver1.07.ks.json"
+#file_path = "D:\\games\\sabbat of witch\\output\\data.xp3\\scn\\005.共通－因幡めぐる.ks.json"
 folder_path = "D:\\games\\senrenbanka\\outPath\\data.xp3\\scn"
+#folder_path = "D:\\games\\sabbat of witch\\output\\data.xp3\\scn"
 
 datas = Scenes(file_path)
-folder = Scnfolder(folder_path,debug=False)
-
+folder = Scnfolder(folder_path)
 
 for data in datas:
-    print(f"{data}")
+    print(f"{data},{type(data)}")
 print('')
 
 for data in datas:
@@ -27,5 +30,5 @@ for file in folder:
             print(f'        {target._name}')
 print('')
 
-#datas[1].exposeTextWithFilter(watch_output = True)
-#datas.exposeTextWithFilter()
+#datas[3].exposeText(watch_output = True)
+#datas.exposeText(watch_output = True)